# [WIP] [PERF] stock_account:16.0 improve _validate_accounting_entries