The function _validate_accounting_entries is notorious for not using proper batching which results in a lot of SQL queries. Originally, it had to iterate individual svls and invoke functions in an unbatched way, even when all svls belonged to the same product. The changes here are to use as much batching as possible and enforcing the use of prefetch_ids to lower the total number of queries.

Benchmarks:

|no. queries before| no. queries after | total time before | total time after|
|-----|-----|------|-----|
|95991|83352|1.59 min|1.48 min|


opw-4283347
opw-4096108




---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr