@robodoo 2024-11-06
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/186434.png)](https://mergebot.odoo.com/odoo/odoo/pull/186434)