@robodoo 2024-11-18
@mohamedkasem99 @jbw-odoo staging failed: ci/template on 4defbec4b3c6f62e2c805febd09c1976d40618ab (view more at https://runbot.odoo.com/runbot/build/70746104)